pyproject.toml
[tool.poetry]
name = "test-cloud-vision-api-google"
version = "0.1.0"
description = ""
authors = ["Adrien Montagu <adrienmontagu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
google-cloud-vision = "^2.6.2"
webcolors = "^1.11.1"
spicy = "^0.16.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
